+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SysRng` behind new feature `sys_rng` [#751]
 - WASIp3 support [#779]
 - `extern_impl` opt-in backend [#786] [#794]
+- Motor OS support [#797]
 
 ### Changed
 - Use Edition 2024 and MSRV 1.85 [#749]