+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2026-02-03
+
+### Fixed
+- Documentation build on docs.rs [#801]
+
+[#801]: https://github.com/rust-random/getrandom/pull/801
+
 ## [0.4.0] - 2026-02-02
 
 ### Added
@@ -659,7 +666,8 @@ Publish initial implementation.
 ## [0.0.0] - 2019-01-19
 Publish an empty template library.
 
-[Unreleased]: https://github.com/rust-random/getrandom/compare/v0.3.4...master
+[0.4.1]: https://github.com/rust-random/getrandom/compare/v0.4.0...v0.4.1
+[0.4.0]: https://github.com/rust-random/getrandom/compare/v0.3.4...v0.4.0
 [0.3.4]: https://github.com/rust-random/getrandom/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/rust-random/getrandom/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/rust-random/getrandom/compare/v0.3.1...v0.3.2